+++ src/assets/scss/forms.scss
@@ -197,6 +197,7 @@
     .btn-icon {
       position: absolute;
       right: 0px;
+      top: 0px;
     }
   }
 }
